+++ src/main/java/com/webshop/internship/service/OrderService.java
@@ -1,7 +1,7 @@
 package com.webshop.internship.service;
 
-import com.webshop.internship.controller.OrderAPIController;
 import com.webshop.internship.model.Order;
+import com.webshop.internship.model.OrderForm;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
@@ -13,7 +13,7 @@ public interface OrderService {
 
     Order create(@NotNull(message = "The order cannot be null.") @Valid Order order);
 
-    Order prepare(OrderAPIController.OrderForm form);
+    Order prepare(OrderForm form);
 
     void update(@NotNull(message = "The order cannot be null.") @Valid Order order);
